main.py

from slither.slither import Slither

def parse_solidity_file(file_path):
    """
    Parses a Solidity file using Slither and returns the parsed object.
    """
    slither = Slither(file_path)
    return slither




def analyze_contract(file_path):
    """
    Runs all gas inefficiency checks on a Solidity file.
    """
    slither = parse_solidity_file(file_path)

    analysis_results = {}

    print("Running gas optimization checks...")

    # analysis_results["detect_something"] = detect_something(slither)

    return analysis_results


def generate_report(results, output_file="report.json"):
    """
    Generates a JSON report of the analysis results.
    """
    import json
    with open(output_file, 'w') as f:
        json.dump(results, f, indent=4)
    print(f"Report saved to {output_file}")


if __name__ == "__main__":
    import argparse

    parser = argparse.ArgumentParser(description="Static Analysis Tool for Gas Optimization in Solidity Smart Contracts")
    parser.add_argument("file", help="Path to the Solidity file")
    parser.add_argument("--output", default="report.json", help="Output file for the report")

    args = parser.parse_args()

    results = analyze_contract(args.file)
    generate_report(results, args.output)
