requirements.txt
slither-analyzer==0.9.4
solc-select==0.2.5